'ft_strlen', change type of loop variable 'i' from int to size_t and 'ft_strrchr', change return value from 0 to NULL

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jfidalgo <jfidalgo@student.42barcel.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/20 20:06:50 by jfidalgo          #+#    #+#             */
-/*   Updated: 2024/02/01 16:03:15 by jfidalgo         ###   ########.fr       */
+/*   Updated: 2024/02/07 14:53:43 by jfidalgo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
